Unused validator
As per checkit docs. Modified by tkellen, and again modified by me.

Instead of having to have knex in scope, I thought it would be nice to pass it in as a param. Because of this, the rule cannot be called using

```javascript
email: ['unused:tablename:columnname']
```
But must be used like this:
```javascript
email: {
  rule:'unused',
  params:['users', 'email', knex]
}
```
Thoughts on the injection of knex?